Add util function ParseVariadicArguments
This utility function is used to parse the cli arguments
for creating service properties.

A service in this context is a named collection of attributes,
with attributes being key/value pairs with predefined keys.
The possible attributes are:
- transport, for the transport protocol
- application, for the application protocol
- port, for port numbers
- process, for process names
- file, for related files paths
- directory, for related directories paths
- socket, for related socket paths
- tls, for specifying tls
- uid, for the process uid
- provider, for the upsteam software provider

Of these, only the `tls` and `provider` attributes are limited to
being given once; all other attributes can be specified multiple
times. This leads to complicated command line specifications, here
also with a team attribute since services are created per team (team
is also limited to once, but not an attribute):

./somaadm property create service 'DNS Resolver [unbound]' \
        team 'xxx' \
        transport tcp \
        transport udp \
        port 53 \
        application DNS \
        process unbound \
        uid 59 \
        file /etc/unbound/unbound.conf \
        file /var/run/unbound.pid \
        directory /var/unbound \
        provider "NLnet Labs"

The cli command framework consumes 'property create service' and
'DNS Resolver [unbound]' is the name of the service that is to be
created. The variable arguments after that are parsed by this
function into a map[string][]string (string map of slices of strings).

The function differentiates between keys that are to be specified:
- at least once
- at most once
- an arbitrary number of times

These constraints are checked; the function aborts upon violation.