BUGFIX: runtime type coalescion panic
Both SomaTreeElemBucket and SomeTreeElemGroup pass .Receive()
requests down the tree if the Receiver is not a direct child
of them; both can have children that can have children of their
own.

But both of these can also have SomaTreeElemNode as child, which
is not a SomeTreeReceiver since it is a pure leaf.

Prevent the fallthrough routine from accessing SomaTreeElemNode
via .(SomaTreeReceiver) interface => panic!