Fix: do not include team as service attribute
somaadm erroneously included the team specification as a service
attribute when attempting to create a service. `team` as a keyword
is explicitly added as allowed for services, meaning it does not
trigger unknown keyword errors from cli parsing based on the attribute
list received from SOMA.

But since `team` is not registered as a service attribute on the server
side, the service can not be created.

This fixes the following error:

% ./somaadm property create service unbound \
        transport_protocol tcp \
        transport_protocol udp \
        port 53 \
        process_comm unbound \
        process_args '/usr/sbin/unbound' \
        file_path '/etc/unbound/unbound.conf' \
        team 'xxx' | jq .
{
  "code": 500,
  "status": "ERROR",
  "text": [
    "pq: insert or update on table \"team_service_property_values\" violates foreign key constraint \"team_service_property_values_service_property_attribute_fkey\""
  ]
}