Fix: use new *_system_properties column source_type
Switch from inserting the source_type into column object_type
to new column source_type. Use the fact that object_type has now
a default value appropriate for the table and does not need to be
set, ie. no change in statement parameter count.

This fixes the following error:
        pq: insert or update on table "node_system_properties" violates foreign key constraint "node_system_properties_system_property_fkey1"

The system property validity check needs to be defined on the object
type of the element that has the property instance, not the object type
of the element where it was inherited from.