0.5.3: add table public.schema_versions
Add new table public.schema_versions and with it the concept of
version database schemas. This is in support for the upcoming
`somadbctl upgrade` command which will be increasingly important
as recreating the database from scratch will no longer be a viable
option.

The initial entries currently installed look like this:

soma=> select * from public.schema_versions;
 serial |  schema   |   version    |         created_at         |           description
--------+-----------+--------------+----------------------------+----------------------------------
      1 | inventory | 201605060001 | 2016-05-06 16:45:13.249+02 | Initial create - somadbctl 0.5.3
      2 | auth      | 201605060001 | 2016-05-06 16:45:24.018+02 | Initial create - somadbctl 0.5.3
      3 | soma      | 201605060001 | 2016-05-06 16:45:32.378+02 | Initial create - somadbctl 0.5.3
(3 rows)

Upgrades will insert additional rows, meaning this table also acts as
history table.

The current version for a schema can be queried like this, example for
inventory schema:

        SELECT   version
        FROM     public.schema_versions
        WHERE    schema = 'inventory'
        ORDER BY serial DESC
        LIMIT    1;

The version scheme is numeric with 12 digits. The first 8 are year,
month and day in %Y%m%d strftime(3) format. The next 4 digits represent
a running counter, starting at 1 every day.

More than 9999 schema versions per day are not supported. This limit is
expected to be irrelevant in practice.

The table is defined as follows:

                                           Table "public.schema_versions"
   Column    |            Type             |                                Modifiers
-------------+-----------------------------+-------------------------------------------------------------------------
 serial      | bigint                      | not null default nextval('public.schema_versions_serial_seq'::regclass)
 schema      | character varying(16)       | not null
 version     | numeric(16,0)               | not null
 created_at  | timestamp(3) with time zone | not null default (now())::timestamp(3) with time zone
 description | text                        | not null
Indexes:
    "schema_versions_pkey" PRIMARY KEY, btree (serial)