Fix: set ConfigId for check instances
When constructing a check instance, the ConfigId field was not set,
leading to a nil UUID default value.

During checkInstance.MakeAction() this nil UUID propageted into the
action channel for TreeKeeper to act on. Since no check configuration
for the nil UUID had been previously saved, this lead to the following
SQL error trying to store the update:

pq: insert or update on table "check_instances" violates
    foreign key constraint "check_instances_check_configuration_id_fkey"

The error was encountered for a node. Fix is for nodes, clusters and
groups (bug proliferation via copy/paste).