tkLoader: reusable treeKeeper.orderGroups()
Add interface tkLoader and convert orderGroups() to accept that
interface instead of struct *tkLoaderChecks directly.

Property loading for groups currently does not respect the order
of groups; but once properties can be overwritten further down in
the tree, ordering becomes vital for them.
Since 5+ future users of this function are already known, make it
reusable right from the beginning.