Partial support for loading CheckInstances from DB
Tree elements group, cluster and node received a new field
loadedInstances, map[string]map[string]CheckInstance. This maps
from CheckId to InstanceId to CheckInstance.

Interface Checker received a new method LoadInstance(i CheckInstance)
which can be used to write a CheckInstance into loadedInstances{}.
On fault, repository and bucket elements this is a no-op, they can
not have instances.
This method is not passed along the Tree, it has to be called on the
correct element, ie. after a FindRequest().

Interface Attacher received a new method ClearLoadInfo() that ensures
that loadedInterfaces{} is empty. This is implemented all the way to
SomaTree and can thus be called as tk.tree.ClearLoadInfo().
It is passed down the tree.

Group.updateCheckInstances() has been modified to switch into startup
mode if loadedInstances{} is not empty. It then tries to match all
computed instances to loaded instances. Any leftovers, whatever direction,
are an error. Well, a panic for now.
During startup, the matched computed instance keeps the InstanceConfigId
and instance Version from the matched loaded instance.

Similar adjustments must be made to Cluster and Node. These will come in
a followup commit.

Faults during startup should utilize the errorchannel and trigger a broken
TreeKeeper.