FIX: commit fixed load-check-instance algorithm
This commit is again very messy, will be cleaned up in later commits.

The new version of the algorithm has so far only been tested in a
repository with a single check on the repository itself, but that
was one of the most wrong cases in the old one.
All computed instances appear to have been matched, no leftover instances
dangling. All createNew actions were suppressed. Analysis is not yet
finished, but this is already much much better than the previous version
that was wrong all over the place.

This update fixes three huge brainfart-style bugs:

1. Do not use current_instance_config_id to load the check instance
   configuration. The name is tempting, but this is the currently deployed
   version. There can be any number of newer versions currently blocked
   within the rollout workflow.

   The tree is supposed to have the actually latest. For now, use the
   created timestamp until the reliability of the version column has been
   explored some more.
   Alternatively, switch to a serial column with an automatically
   maintained sequence.

2. Do not reuse the lists of check IDs used to assembly checks to process
   check instances.

   The used check IDs were checks defined on the tree element type that
   were source checks, ie. not inherited. This is of course an insufficient
   subset of checks to work on for loading instances.
   It is required to also load instances for checks that were inherited.

   Using the wrong check ID list lead to large gaps in the loaded instances,
   explaining all the previously unmatchable excess instances.

3. If a type has no source check on it, do not return and continue with the
   next type. Instead, only skip over check assembly and go directly to
   loading check instances, which can not be skipped over.