Add Kex.SetIPAddressString(s string) method
This is the more dangerous cousin to Kex.SetIPAddress(http.Request)
which skipped input validation since it accesses the http.Request
directly for the remote address string.

SetIPAddressString still requires the input string to be in the
format of http.Request.RemoteAddr. It should only be used if it is
certain that the string represents a RemoteAddr, set by a trusted
caller.