Implement supervisor.Validate()
This function can be used during http basic auth to
validate (account, token, srcAddr) tuples.

The commit includes supervisor.fetchTokenFromDB() which
can be used to load a single token from the database.
This is useful for readonly instances that have to check
the DB if the readwrite instance created the token.