Update constraint accounting for @defined
Update all constraint collectors to return the value they bound
against. Without @defined, the user congigured value could be used as
the bindings were exact matches.
Now `constraint attribute port @defined` would match 80, and after the
service is changed 443. But the ValueHash would be computed over
@defined.
With no change, this update would be eliminated, instead of triggering
a new rollout.

While here, fix a bug that caused evalAttributeOfService to always
return false. Before e1af0820c8c15142baeae772f5e26c958309b4f9, the
first parameter passed to it was the view of the check (yes, twice).
Since then, it is the Id of the service property.
It was and is not the service name however, like the signature
suggested, and what it was compared to - with very little success.