Make timestamps deterministic for hash computation
To have a stable sub-second part of the Go time.Time value,
it needs to be exported/imported using time.RFC3339Nano
otherwise there are fluctuations.

Since PostgreSQL can only store Microsecond precision, this
is not possible. Nanosecond precision token expiration is also
a bit silly.

To make the timestamps deterministic, after marshalling them
to binary []byte, the 4 bytes representing the subsecond
resolution are overwritten with a fixed value 0xdeadcafe.