FALLOUT WARNING! Convert calls to new Client
So far, every call opened its own single-use resty
Client. With https, and the authentication setup, this
would be a huge pain.

Convert all calls to being passed in a *resty.Client as
part of the signature, reusing one Client instance.

This also makes it easier to write unit tests, if someone
would write them.