NodePropertyAdd: improve node-not-assigned error
To attach a property to a node, that node must be assigned to a
configuration repository. If the node has not been assigned yet,
the error so far was a generic 404 Not Found.

That this really meant `could not find a repository the node belongs
to` was non-obvious.

Improve the clientside lookup logic to provide a better, more useful
error message.