0.7.39: FIX token generation for some source ip
The remote address was pushed through ExtractAddress twice, which
is not idempotent for some addresses, for example ::1.

This caused the resulting IPAddr to be empty. Token generation
checks that the IP mixed into the HMAC is not equal to the empty
object. This caused the following fault:

Subsystem=supervisor, Action=bootstrap_root, InternalCode=500, Error=Invalid input
Subsystem=supervisor, Request=bootstrap_root, InternalCode=500, ExternalCode=401

Constrol resulting fallout and check all instances where this SourceIP
is used during KEX or other actions.