Handle updates for properties with overwrites
If an inherited property has receive a local overwrite,
the object still receives updates from up top.

If we can detect to have an overwrite, ignore the update
without error and do not pass it further down.

If the property to update can not be found and no overwrite
exists, propagate an error to the Fault Element.