Handle duplicate setProperty inherited from above
The reverse overwrite case, an object inherits a SetProperty
for a property it already has and is the source instance for
it.

This occurs if the property is set a second time further up in
the tree.

Handle it by stopping the inheritance and keeping the locally
set property.

If the duplicate SetProperty is for a property that the source
instance is not local, then the tree is corrupted. We should
be the source, or should inherit an update instead of a set.
Abort via Fault element.

Include testcase: TestSetAboveSetProperty