Handle updates with inheritance changes
It is possible to update a property without inheritance and
setting it for the update. Likewise, a property with inheritance
can be updated to no inheritance.

In both cases, updatePropertyOnChildren can not be used. The
children either do not have the property to update or must delete
the old property.

Add handling to switchProperties to detect inheritance changes.

For true->false transitions:
- call deletePropertyOnChildren(oldProperty)

For false->true transitions:
- call setPropertyOnChildren(newProperty)

Testcases:
- TestUpdatePropertyInheritanceFalseToTrue
- TestUpdatePropertyInheritanceTrueToFalse