v0.7.46: fix nilptr dereference in deployments
 /usr/local/go/src/runtime/panic.go:464 +0x3e6
 xxx/xxx/soma/vendor/xxx/xxx/somaproto.(*Capability).DeepCompare(0x0, 0x0, 0x0)
 /home/xxx/go/src/xxx/xxx/soma/vendor/xxx/xxx/somaproto/capability.go:33 +0x821
 xxx/xxx/soma/vendor/xxx/xxx/somaproto.(*Deployment).DeepCompare(0xc83d77f3b0, 0xc83d77f2c0, 0xd00)
 /home/xxx/go/src/xxx/xxx/soma/vendor/xxx/xxx/somaproto/deployment.go:52 +0x40b

This nilptr derefence means that during the deep compare either the current or previous
deployment details missed the capability section. It can not be the current one, since
then the nilptr deref would not be inside the DeepCompare call but the DeepCompare call
itself.

It is unclear what created deployment details without capability section. But if the
current version has them, then we simply ignore the missing previous section since this
deployment calculation will fix the details.