Implement handling for special disable properties
This commit adds support for two special-cased system properties:
- disable_all_monitoring
- disable_check_configuration

The first property, if set on the object, removes all existing
check instances that are created in the same view as the property.
To remove all internal checks:
        disable_all_monitoring=true, view=internal
To remove all checks:
        disable_all_monitoring=true, view=any

The second property, if set on the object, removes all existing check
instances for checks that were created by a specific check
configuration. The view of the property has to match the view of the
check configuration though.

To accomplish nothing:
- create checkconfig b1de01c1-98fd-47f5-899f-a8ccc43909cb
  in view internal
- set disable_check_configuration=b1de01c1-98fd-47f5-899f-a8ccc43909cb
  in view external

To disable it regardless of view (recommanded):
  disable_check_configuration=b1de01c1-98fd-47f5-899f-a8ccc43909cb,
  view=any

Since these are regular system properties, the usual inheritance rules
apply. This means they can be used to black out either a single object
or an entire subtree hierarchy.