Cleanup soma operation modes; default environment
The default environment if not specified is `production`.

Modes:
- readonly:
  * start DB read handlers
  * does not load trees
  * does not load job handler
  * does not load lifeCycle management
  * supervisor runs but issues no credentials
  ==> this mode is intended to potentially run
      multiple times against one database, does
      not need INSERT/DELETE/UPDATE db access

- observer:
  * start DB read handlers
  * trees are loaded, but no deploymentdetails
    maintenance
  * lifeCycle management disabled
  * job handling disabled
  * supervisor runs and issues only new tokens
  ==> this mode is basically read-only-with-trees

- normal:
  * full r/w instance with all capabilities

For both observer and normal mode it is possible to only
load a single tree instead of all, using the `-repo ${name}`
command line argument.
In normal mode, this is only allowed if the config
file environment is set to something other than
`production`, the default value. It can be forced with
the -allowdatacorruption flag.

If multiple r/w instances, single-repo of full, against a
single DB lead to a corrupt system, support will not
extend beyond a good laugh and a pat on the back.