FIX: no multiple init of tracking datastructure
ckOrder is of type and map[string]map[string]tree.Check, being filled
in the following way:
        objId -> checkId -> tree.Check

This allows to traverse checks by object. Of course one object can
have any number of checks on it, so reinitializing with an empty
checkId-map loses items, which are then not loaded into the tree
and can not be updated/deleted due to this.