WIP: factor out transaction initialization
Starting the transaction that is used to process the
action channel events and setting up all the prepared
statements is a lot of mundane declarations, that
distract from the codeflow in TreeKeeper.

Factor them out into a function that also returns a map
to all the prepared statements.

As of this commit, this branch does currently not
build. This will be the case until the actionloop has
been fully taken apart.