Vendor import: go-sqlmock
The soma/main.treeKeeper.process() method is being refactored in a
different branch. It is the central method for asynchronous tree
manipulation requests.
To write unit tests for the current version that can then be used to
test the refactoring against, a mock-up database lookalive is required,
since the method is very SQL statement heavy.

For this reason, vendor-import DATA-DOG's go-sqlmock to use it in those
tests.
Update Copyright file to acknowledge the vendor redistribution.