Refactor guidepost/3: data fixing, late validation
- move all data fixing into methods with a top-level
  interface that does the right thing

- expand the routing id extraction to make it more
  general purpose.

- add handler validation based on the improved extraction

- rework the late input validations that happend after the
  handler validation to be part of the first, now single,
  input data validation pass

While here, clean up the error path. Since it would no longer
be possible to return a 404/NotFound since sql.ErrNoRows does
not contain information which lookup failed, introduce a notFound
flag that is propagated along the error message, to indicate if
the error should be a ServerError or NotFound.