TreeKeeper: abort jobs if setting start mark fails
If the SQL statement that marks the job as being processed
fails, then completely skip the job.

The likely results from the previous handling was that a later
SQL statement also failed, resulting in an aborted job; rolling
back the in-memory tree.

The unlikely case would have been a successfully completed job
with no start-time for the job (only queued and finished).