FIX: ignore deleted checks as source check
When looking for the source-/root check for a check configuration,
do not select an already deleted check for this check configuration
as the loaded tree will not match it.

This can happen if either the repository had a checks-level rebuild
or a property such as disable_all_monitoring was used.