Treekeeper: remove log.Fatal from runtime path
log.Fatal() is used throughout when a SQL statement fails
to prepare, because there is no way to recover from a syntax
error in the statement or something similar.

Since statements are usually prepared during startup, this is
desired behaviour.

Treekeeper goroutines differ from this, since they are also
created at runtime, either new due to a create_repository request
or simply because the treekeeper for an existing, running
repository was restarted.

Instead of aborting the service, mark the tree as broken and
bail.