Replace stmt variable to avoid scope problems
Use `statement` instead of `stmt`, since stmt is also
the name of an imported package. This could lead to
scope problems later on.

They are caught as compile time errors (use of package
stmt without a selector), but there is no reason to
not use a different var and improve readability.