Fixes for go vet
config.go:
- malformatted json tag for PokeBatchSize
- unexported url has json tag declaration

logrotate.go:
- format string in Print statement

soma_guide_post__validate.go:
- no format string in Errorf statement

soma_handler_tree_keeper__order.go:
- format string in Println statement

soma_handler_supervisor.go:
This file is changed to clean the `go vet` output. Generally, it is
not allowed to copy a lock by value. It is however safe to copy a
zero value lock by value, as a special case and the only one that does
not lead to strange results.
This is what was done here, the locks were copied in the return of the
.newXXX() functions that allocated them; this was not a bug. But it is
a weird special case that I also only learned by consulting the great
tome of google in reference to this error output.
Issue https://github.com/golang/go/issues/13675 explains the zero value
special case.
To avoid this confusion, make the functions return pointers and update
the resulting types.