Lifecycle: convert prepared SQL statements in tx
For transactions, the range loop idiom that other
functions were converted to does not work. In that
construct the statement is actually prepared against
a copy of what is later used, but since it is a copy
of a pointer this is fine.

The only thing that actually uses this pointer copy
though is the defered call to .Close() normally issued.
In Tx context, this call is not required, since
statements are closed by either .Commit() or .Rollback().

A double .Close() may cause up to nilptr-deref panics.
This means the declared copy is never used, causing
`go build` to refuse the compile.

Convert to the way it is done in soma_tree_keeper__tx_sql.go,
looping over map[string]string and then preparing into
map[string]*sql.Stmt.
Convert later uses of the prepared statements to use the
map.