Tree: update marker for groups, clusters and nodes
Add an update marker to the three object types that
can have check instances. The flag is set to true on:
- addProperty() / rmProperty()
- addCheck() / rmCheck()

If the marker is not set, then the object has not been
updated in a way that is relevant to check instances.
In that case, updateCheckInstances() can bail early
without sending out update actions.

This results in fewer new check instance configurations
that have to be computed, compared to the previous version
and then deleted without rollout since nothing changed.