Add functions to export proto.CheckConfig from DB
exportCheckConfig() and exportCheckConfigThresholds() are
the first functions that encapsulate building specific
structs from the database.

The goal is of course to be able to reuse these functions
in different handlers. Since handlers can have their own
connection pool to the database, these functions are
`bring your own prepared statement`.

This of course also allows to use them as-is, or hand them
statements that were prepared inside a transaction.