Add export function exportCheckConfigObjectTX()
This function encapsulates all other check config
export functions in a transaction context and returns
a slice pointer containing all the check configurations
that are active on a specific object, with all the
instances they have created.

The function commits the transaction if no error is
encountered; the caller has to roll it back on error.