fc.spawnTreeKeeper: properly pass errors along
Ever since the introduction of per-TreeKeeper database pools
this method could actually fail. But the signature was not updated
at the time, and errors dropped.

The impact of this was rather low though, since I database connection
problem would manifest many other errors.

This was forgotten about and then rediscovered during the change to
per-TreeKeeper logfiles, when the errors on file.Open could not be
passed along. This is a much more likely error case for long running
processes, so this needed to get fixed and was flagged as such
in the local tree.

Update the signature and pass errors fully up, including request
failures to the client.