Convert package tree to non-global logger
TreeKeeper routines were switched to separate
logfiles, but the in-memory tree was still using
the global logger.

Add support to the tree to receive a local logger
instance that is propagated down.

Update forestCustodian to create the trees with
a local logger.

Update treeKeeper to switch the tree first to the
startup and then the regular per-tree log.