Deployment: add task `delete`
The only two values for deployment.Task so far were
`rollout` and `deprovision`. This commit adds a third
task `delete` that can be used instead of `deprovision`.

The decision which task to give out is based on if there
are updated configurations blocked on the rollout. If
there are newer versions the task is `deprovision`.

If there are no newer versions and the check instance is
fully removed, then the task is `delete`.

This allows to differentiate between remove with or without
alarm clearing for the deployment consumer.