adm: limited cli argument validation
For commands with one or zero arguments, the validation
based on argument count is required, since the argument
is usually a runtime value.

No generic number-of-arguments function is provided as
can be found in package util.
The Verify* functions return error instead of aborting
as found in util.

This commit adds:
- VerifySingleArgument()
        verify there is exactly one argument
- VerifyNoArgument()
        verify there are no arguments
- AllArguments()
        a small helper to extract all args from context