somaadm: update arg count based validation
For commands with more than 1 argument, argument based
validation is roughly useless since it does not report
what is missing.
Since ParseVariadicArguments does report that, just skip
the count validation and input directly into it.

Convert all commands with more than 1 argument that did
not use ParseVariadicArguments to use it.

For zero/one argument commands switch to the new functions
in package adm.