Fix return value passing
Maps are pointer-values, so the copy of result given to
the function is a copy of the caller's pointer.

Setting it to an empty map only updates the functions
pointer to a new map. This will then be filled with the
function results and garbage collected once the function
returns and it goes out of scope.
The map passed in by the caller will stay empty.

Found while adding ParseVariadicCheckArguments to the file.