soma: http/tree, update authorization check
This commit removes support for the tree command on
nodes. It was always rather pointless to request a
recursive view of the children of a leaf object that
not have any.

Now that this would also be a namespacing authorization
problem, there is just no point to having it around.