perm: start objectLookup for basic repo tracking
This cache object will not reimplement the tree, but only
track to which repository/bucket an object belongs. These
are the two initially supported objects to grant permissions
on.

When full-depth permissions are implemented, this module will
stay the same, but switch to answering `100, continue` over
`200, ok`. The final permission decision will then be implemented
by the tree itself.