perm: restore performPermissionAdd
performPermissionAdd() was removed along with
performCategoryAdd as these would be implicitly
tracked instead. But with both being implicitly
tracked, the information to which category a permission
belongs is no longer covered anywhere.
This in turn would mean the rmCategory case can not
determine which permissions are in that category and
must be removed.

Track permission->category relations explicitly in
permissionMapping.