Fix: clear check UUID before inheriting it down
If the UUID of the passed down check is not uuid.Nil,
the child will not generate a new UUID for its check
causing duplicate key errors when trying to persist
the tree update.

While here, go the extra mile and clone the check to
ensure the child gets its own copy with no dangling
pointers to the parents' check.