soma: rework startup via Handler interface
This commit adds functions that return the handlers
that have been migrated so far:
        - Node Read/Write
        - Validity Read/Write
        - View Read/Write
The length of the input channel buffer is now a configuration
setting.

All handlers also received a register() function to pass them
resources that are initialized by the SOMA application.

A Handler interface has been defined that contains the required
methods of a Handler. HandlerMap has been updated and uses the
Handler interface instead of interface{} as value type.

This allows to shrink Soma.Start() by using a range loop over the
HandlerMap instead of repetitive startup functions for every
handler.